driver-service/requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
sqlalchemy==2.0.24
databases==0.9.0
httpx==0.28.1
aiosqlite==0.21.0
python-dotenv==1.0.1
pydantic==2.11.9
pydantic_core==2.33.2
typing_extensions==4.15.0
annotated-types==0.7.0
colorama==0.4.6
click==8.3.0
starlette==0.48.0
sniffio==1.3.1
anyio==4.11.0
typing-inspection==0.4.2

# AWS async dependencies
boto3
aioboto3
botocore
asyncpg
psycopg2-binary
prometheus_client
python-jose[cryptography]

